Split into IBMSemeruTest and Java8to11Test
